remove access_token from installation params
