+++ apps/typeform/frontend/src/constants.ts
@@ -1,0 +1,1 @@
+export const SDK_WINDOW_HEIGHT = 450;
