+++ apps/typeform/frontend/src/index.html
@@ -2,6 +2,7 @@
 <html>
   <head>
     <meta charset="utf-8" />
+    <title>Typeform</title>
   </head>
   <body>
     <div id="root"></div>
